feat: change output checking from text-file based to pipe based
